various TODO markers, implement RenderLayoutPlugin and RenderModelPlugin

+++ extensions/render/src/org/sbml/jsbml/ext/render/Ellipse.java
@@ -31,5 +31,6 @@ package org.sbml.jsbml.ext.render;
  * @date 08.05.2012
  */
 public class Ellipse extends GraphicalPrimitive2D {
-	protected Integer cx, cy, cz, rx, ry; 
+	protected Double cx, cy, cz, rx, ry; 
+	//FIXME absolute booleans
 }
